server/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
deepface==0.0.83
python-dotenv==1.0.1
werkzeug==2.3.8
langchain==0.1.16
langchain-groq==0.1.1
tensorflow==2.12.0
keras==2.12.0
requests==2.31.0
gunicorn==21.2.0
